refactor: consolidate user schema imports into a single schema file
