+++ src/app/users/users-data.ts
@@ -3,7 +3,7 @@ import 'server-only';
 import { cache } from 'react';
 import { eq, desc, asc, sql, SQL } from 'drizzle-orm';
 import { db } from '@/db';
-import { users } from '@/db/schema/users';
+import { users } from '@/db/schema';
 import { type Role } from '@/lib/types';
 
 export interface GetUserResponse {
